feat(资源评分): 修改评分字段类型为double并优化API设计
将资源评分的recommendation字段从int64改为double类型以支持更精确的评分
重构SubmitResourceRating API，使用resource_id作为路径参数
移除不必要的rating_id字段，简化请求结构
更新相关路由和模型定义以匹配新的API设计

+++ idl/model.thrift
@@ -90,7 +90,7 @@ struct ResourceRating{
     required i64 ratingId,
     required i64 userId,
     required i64 resourceId,
-    required i64 recommendation,
+    required double recommendation,
     required bool isVisible,
     required i64 createdAt,
 }
